requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-oauthlib==1.2.1
google-cloud-firestore==2.20.2
pydantic==2.11.3
pydantic-settings==2.8.1
python-dateutil==2.9.0.post0
dateparser==1.2.1
